examples/requirements.txt
# Core asynchronous library
anyio>=4.5

colorama

python-dotenv

graphviz

groq
openai>=1.0.0 

httpx>=0.27
httpx-sse>=0.4

mcp>=1.2.0

# JSON Schema validation for tools
jsonschema

requests

crawl4ai

duckduckgo-search
pydantic>=2.7.2,<3.0.0
pydantic-settings>=2.5.2 

